requirements/base.txt
Django==3.2.5
django-auth-ldap==2.2.0
django-compressor==2.4
django-crispy-forms==1.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-reversion==3.0.8
django-tables2==2.3.3
django-tables2-column-shifter==0.5.2
django-widget-tweaks==1.4.8
openpyxl==3.0.5
python-ldap==3.3.1
pytz==2020.4
six==1.15.0
tablib==3.0.0
Werkzeug==1.0.1
xlrd==2.0.1
xlwt==1.3.0
